requirements.txt
numpy~=2.1.3
tensorflow~=2.19.0
pillow~=11.2.1
ollama~=0.4.9